+++ Lawn/ZenGarden.cpp
@@ -2488,7 +2488,7 @@ SeedType ZenGarden::PickRandomSeedType()
             aSeedCount++;
         }
     }
-    return (SeedType)TodPickFromArray((int*)aSeedList, aSeedCount);
+    return (SeedType)TodPickFromArray((intptr_t*)aSeedList, aSeedCount);
 }
 
 //0x5223F0
